adjusted file structure
