+++ package.json
@@ -27,9 +27,11 @@
   },
   "homepage": "https://github.com/AllInOpenSource/Project-Badging#readme",
   "dependencies": {
+    "final-form": "4.20.9",
     "prop-types": "15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-final-form": "6.5.9",
     "react-router-dom": "^6.10.0",
     "react-scripts": "^5.0.1"
   },
