+++ src/components/index.js
@@ -3,5 +3,5 @@ import Footer from './Footer/Footer';
 import Faq from './Faq/Faq';
 import CardComponent from './HowItWorks/Card';
 import CardCarousel from './Carousel/CardCarousel';
-import BadgingForm from './BadgingForm/BadgingForm';
-export { Header, Footer, Faq, CardComponent, CardCarousel, BadgingForm };
+
+export { Header, Footer, Faq, CardComponent, CardCarousel };
